Use the same test script for step/next tests
Merging the tests together with the same program makes it easier to
observe the difference between commands.
